make large number check remember all digits of a float
When a float is converted from source to an internal representation
trailing 0s in the fractional part are lost.  If the programmer
deliberately used a number like 1_234.567000 in the code then we should
not complain because tokenising "loses" the trailing 000.

This also contains a workaround for a bug in Elixir prior to 1.3.2 where
the embedded _ in numbers was not accounted for in the columns provided
in the location tuple returned from `:elixir_tokenizer.tokenize/3`.  The
workaround is not that good because once the columns are out of sync
then a completely correct solution would have to re-tokenise the line
(or more).