stop long number readability check complaining about non-decimals
The rule of _ every three digits makes sense for decimal digits, but
a mesasge suggesting 0xFFFF would be better written as 65_535 is not
helpful.

This patch lets the programmer format binary, octal, and hexadecimal
numbers in any way in the source.  I am not sure there is a "right" way
of grouping non-decimal numbers because the context might affect the
logical way to split non-decimal numbers... in this case leave it to the
programmer.