Add ability to use `--no-strict` command line flag
We implemented the ability to set a `strict: true` config in the `.credo.exs`
file for a project, but when we did that we didn't implement any way to
override that with a command line flag. In this commit we're adding that
ability to override that setting with a `--no-strict` flag so you don't need to
edit that file to turn off `strict` mode for just a single run.