Includes the 'only_greater_than' value in the LargeNumbers message
More specificity is helpful in the messages. While the old message gives you the correct value for the violation, it did not offer guidance on avoiding the error subsequently. Additionally, given that the `only_greater_than` is user-configurable, teams that change the default would have to go digging to find the threshold.